Move slugify and filename builder into filenames module
